feat: move interfaces and types to separate file
