+++ Tanks-1990/src/app.ts
@@ -1,3 +1,4 @@
+import './style.scss';
 // @ts-ignore
 import main from "./controller/controller.ts";
